[PCB] Connect RTS pin to Teensy
The original design (v 2.0) accidentally omitted an RTSpin on the Teensy.
Here I added it into the PCB and made associated changes in the Teensy
code. RTS is on Teensy pin 28, and the rest of the Teensy pin map has
changed accordingly. This pin was chosen to:
- place RTS on the easily accessible pins on the side of the Teensy
- not use up a PWM pin, to keep those for valve driving

In rewiring the PCB, I also added more space around the top left standoff.

This fixes #1 and #4.